requirements.txt
requests
python-dotenv
aiohttp
groq
openai
